Major Updates to Grapevine
This was a significant update in that the reformatting of Grapevine HTML has proven inefficient. Rather than rebuilding the grapevine piece by piece, the current standard is to feature a card with a link to the current grapevine (published by membership dept.) accompanied by the first paragraph of the host's message and image. This does not fully incoporate the grapevine into the member communities framework but it creates less room for error.